File structure + push new notifications on top
