+++ Src/Generic/Extensions/UIView/UIView+Viewable.swift
@@ -142,16 +142,17 @@ extension UIView {
 extension UIStackView {
     // swiftlint:disable large_tuple
     private func materializeArrangedViewable<V: Viewable, MatterView: UIView>(
-        viewable: V
+        viewable: V,
+        atIndex: Int?
     ) -> (V.Matter, V.Result, Disposable) where V.Matter == MatterView, V.Events == ViewableEvents {
         let wasAddedCallbacker = Callbacker<Void>()
         let viewableEvents = ViewableEvents(
             wasAddedCallbacker: wasAddedCallbacker
         )
         let (matter, result) = viewable.materialize(events: viewableEvents)
-
-        addArrangedSubview(matter)
-
+        
+        atIndex != nil ? insertArrangedSubview(matter, at: atIndex!) : addArrangedSubview(matter)
+        
         wasAddedCallbacker.callAll()
 
         return (matter, result, Disposer {
@@ -163,12 +164,13 @@ extension UIStackView {
 
     func addArranged<V: Viewable, MatterView: UIView>(
         _ viewable: V,
+        atIndex: Int? = nil,
         onCreate: (_ view: V.Matter) -> Void = defaultOnCreateClosure
     ) -> V.Result where
         V.Matter == MatterView,
         V.Result == Disposable,
         V.Events == ViewableEvents {
-        let (matter, result, disposable) = materializeArrangedViewable(viewable: viewable)
+        let (matter, result, disposable) = materializeArrangedViewable(viewable: viewable, atIndex: atIndex)
 
         onCreate(matter)
 
@@ -180,6 +182,7 @@ extension UIStackView {
 
     func addArranged<V: Viewable, Matter: Viewable, View: UIView>(
         _ viewable: V,
+        atIndex: Int? = nil,
         onCreate: (_ view: Matter.Matter) -> Void = { _ in }
     ) -> V.Result where
         V.Matter == Matter,
@@ -194,7 +197,7 @@ extension UIStackView {
             wasAddedCallbacker: wasAddedCallbacker
         ))
 
-        let (viewableMatter, viewableResult, disposable) = materializeArrangedViewable(viewable: matter)
+        let (viewableMatter, viewableResult, disposable) = materializeArrangedViewable(viewable: matter, atIndex: atIndex)
 
         onCreate(viewableMatter)
